requirements.txt
asgiref==3.8.1
Django==5.0.4
django-ckeditor==6.7.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
pillow==10.3.0
PyMySQL==1.1.1
sqlparse==0.4.4
tzdata==2024.1
